requirements.txt
fastapi==0.109.2
uvicorn==0.23.2
pydantic==2.6.3
python-dotenv==1.0.1
PyJWT==2.10.1
httpx==0.27.2
psycopg2-binary==2.9.9
